pyproject.toml
[project]
name = "adaptive-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "elevenlabs>=2.8.1",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.8",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.5.4",
    "python-dotenv>=1.1.1",
]
